full swagger support&final makeOrder func

+++ internal/models/user.go
@@ -4,7 +4,7 @@ type User struct {
 	Id       int    `json:"id"`
 	Login    string `json:"login" binding:"required"`
 	Password string `json:"password" binding:"required"`
-	Balance  int    `json:"balance"`
+	Balance  float64    `json:"balance"`
 	IsActice bool 	`json:"is_active" db:"is_active"`
 }
